% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankprob.R
\name{convert_rankprob}
\alias{convert_rankprob}
\alias{predict.rankprob}
\title{Convert a multinomial regression to a \code{rankprob} object}
\usage{
convert_rankprob(fit)

\method{predict}{rankprob}(object, newdata = NULL, type = "rank_probs", rankings = NULL, ...)
}
\arguments{
\item{fit}{the output of a multinomial or ordinal regression model, e.g., form \code{nnet::multinom()}, \code{mclogit::mblogit()}, \code{MASS::polr()}, \code{WeightIt::multinom_weightit()}, or \code{WeightIt::ordinal_weightit()}.}

\item{object}{a \code{rankprob} object; the output of a call to \code{convert_rankprob()}.}

\item{newdata}{the new data used to compute predictions. If \code{NULL}, the original data used to fit the model will be used.}

\item{type}{the type of prediction desired. If \code{"rank_probs"}, the probability of the ranking identified in \code{rankings} will be computed. Otherwise passed the \code{predict()} method for the original model fit. Note \code{"rank_probs"} cannot be abbreviated, but as the default, it also doesn't need to be specified.}

\item{rankings}{when \code{type = "rank_probs"}, a list of character vectors that identify the rankings for which predicted probabilities are to be computed. Can also be a single character vector for a single ranking. If \code{NULL}, will compute the ranking of the outcome levels in their original order. Each ranking must contain all levels of the original outcome.}

\item{\dots}{further arguments pass to \code{predict()} when \code{type} is not \code{"rank_probs"}.}
}
\value{
For \code{convert_rankprobs()}, a \code{rankprobs} object, which also has the original model classes. Otherwise the supplied object is unchanged.

For \code{predict()}, a matrix with a row for each unit and a column for each ranking, with each cell equal to the predicted probability of that ranking being chosen for that unit.
}
\description{
\code{convert_rankprob()} converts a supplied multinomial or ordinal regression mdoel to a model tha tsupports computation of probabilities of ranked choices.
}
\details{
\code{convert_rankprobs()} first checks to make sure the model is ordinal or multinomial as determined by \code{\link[insight:model_info]{insight::model_info()}}. Next, it adds \code{"rankprob"} to \code{options("marginaleffects_model_classes")} to facilitate use with \code{marginaleffects} and packages that depend on it, such as \code{clarify}. Finally, it assigned to the supplied object the class \code{"rankprob"}.

\code{predict()} produces the predicted probability of each supplied ranking when \code{type = "rank_probs"}.
}
\note{
When using \code{marginaleffects} functions, \code{type = "rank_probs"} must be specified for the functions to work correctly.
}
\examples{
\dontshow{if (requireNamespace("nnet", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

data <- data.frame(x1 = rnorm(500),
                   x2 = rnorm(500),
                   x3 = rnorm(500),
                   y = sample(LETTERS[1:4], 500, TRUE))

fit <- nnet::multinom(y ~ x1 + x2 + x3, data = data, Hess = TRUE)

fit <- convert_rankprob(fit)

p <- predict(fit, rankings = list(c("A", "B", "C", "D"),
                                  c("A", "C", "B", "D"),
                                  c("B", "A", "D", "C")))

head(p)
\dontshow{\}) # examplesIf}
}
